Steganography_Python/requirements.txt
tk==0.1.0
pillow==11.0.0
numpy==2.2.0